refactor cluster orchestrator
