+++ cluster_orchestrator/cluster-manager/services/analyzing_workers.py
@@ -1,4 +1,4 @@
-from mongodb_client import find_all_nodes
+from clients.mongodb_client import find_all_nodes
 
 
 def looking_for_dead_workers(interval):
