+++ cluster_orchestrator/cluster-manager/services/service_operations.py
@@ -1,10 +1,10 @@
-from cluster_scheduler_requests import scheduler_request_deploy
-from mongodb_client import (
+from ext_requests.cluster_scheduler_requests import scheduler_request_deploy
+from clients.mongodb_client import (
     mongo_create_new_job_instance,
     mongo_find_job_by_system_id,
     mongo_remove_job_instance,
 )
-from mqtt_client import mqtt_publish_edge_delete
+from clients.mqtt_client import mqtt_publish_edge_delete
 
 
 def deploy_service(job, system_job_id, instance_number):
