+++ cluster_orchestrator/cluster-manager/start-up.sh
@@ -3,7 +3,7 @@
 # docker-compose --file ../docker-compose-amd64.yml up -d
 
 # create virtualenv
-virtualenv --clear -p python3.8 .venv
+virtualenv --clear -p python3 .venv
 source .venv/bin/activate
 
 .venv/bin/pip install -r requirements.txt
@@ -15,18 +15,39 @@ export MQTT_BROKER_URL=localhost
 export MQTT_BROKER_PORT=10003
 
 export CLUSTER_MONGO_URL=localhost
-export CLUSTER_MONGO_PORT=10007
+export CLUSTER_MONGO_PORT=10107
 
-# export SYSTEM_MANAGER_URL=131.159.24.210
-# export SYSTEM_MANAGER_URL=3.120.37.66
 export SYSTEM_MANAGER_URL=localhost
 export SYSTEM_MANAGER_PORT=10000
 
 export CLUSTER_SCHEDULER_URL=localhost
-export CLUSTER_SCHUEDLER_PORT=5555
+export CLUSTER_SCHUEDLER_PORT=10105
 
-export CLUSTER_NAME=cluster_thinkpad2
-export CLUSTER_LOCATION=Garching2
+export CLUSTER_SERVICE_MANAGER_ADDR=localhost
+export CLUSTER_SERVICE_MANAGER_PORT=10110
+
+export SYSTEM_MANAGER_GRPC_PORT=50052
+
+# get public IP
+PUBLIC_IP=$(curl -sLf "https://api.ipify.org")
+if [ $? -ne 0 ]; then
+    echo "Error: Failed to retrieve your public IP address."
+    exit 1
+fi
+# get geo coordinates of public IP
+ipLocation=$(curl -sLf "https://ipinfo.io/$PUBLIC_IP/json")
+if [ $? -ne 0 ]; then
+    echo "Error: Failed to retrieve your public IP address."
+    exit 1
+fi
+# Extract latitude and longitude
+latitude=$(echo "$ipLocation" | jq -r '.loc | split(",") | .[0]')
+longitude=$(echo "$ipLocation" | jq -r '.loc | split(",") | .[1]')
+echo Default cluster location $(echo $latitude,$longitude,1000)
+
+export CLUSTER_LOCATION=$(echo $latitude,$longitude,1000)
+export CLUSTER_NAME=cluster_local
+
 export MY_PORT=8000
 
 .venv/bin/python cluster_manager.py
